+++ json_output/cucumber.json
@@ -12,129 +12,163 @@
         "steps": [
           {
             "result": {
-              "duration": 34689538500,
+              "duration": 24092115700,
               "status": "passed"
             },
             "line": 5,
             "name": "user is already on Login Page",
             "match": {
-              "location": "DealsStepDefinition.user_is_already_on_login_page()"
+              "location": "DealsStepWithMapDefinition.user_is_already_on_login_page()"
             },
             "keyword": "Given "
           },
           {
             "result": {
-              "duration": 149732200,
+              "duration": 90695400,
               "status": "passed"
             },
             "line": 6,
             "name": "title of Login Page is FreeCRM",
             "match": {
-              "location": "DealsStepDefinition.title_of_login_page_is_free_CRM()"
+              "location": "DealsStepWithMapDefinition.title_of_login_page_is_free_CRM()"
             },
             "keyword": "When "
           },
           {
             "result": {
-              "duration": 20392669900,
+              "duration": 18680433501,
               "status": "passed"
             },
             "line": 7,
             "name": "user clicks on login button first time",
             "match": {
-              "location": "DealsStepDefinition.user_clicks_on_login_button_first_time()"
+              "location": "DealsStepWithMapDefinition.user_clicks_on_login_button_first_time()"
             },
             "keyword": "Then "
           },
           {
             "result": {
-              "duration": 1434677900,
+              "duration": 2605164200,
               "status": "passed"
             },
             "line": 8,
             "name": "user enters username and password",
             "match": {
-              "location": "DealsStepDefinition.user_enters_username_and_password(DataTable)"
+              "location": "DealsStepWithMapDefinition.user_enters_username_and_password(DataTable)"
             },
             "rows": [
+              {
+                "cells": [
+                  "username",
+                  "password"
+                ],
+                "line": 9
+              },
               {
                 "cells": [
                   "Lokeshnp77@gmail.com",
                   "F!rew0rk"
                 ],
-                "line": 9
+                "line": 10
               }
             ],
             "keyword": "Then "
           },
           {
             "result": {
-              "duration": 173256300,
-              "status": "passed"
-            },
-            "line": 11,
-            "name": "user clicks on login button",
-            "match": {
-              "location": "DealsStepDefinition.user_clicks_on_login_button()"
-            },
-            "keyword": "Then "
-          },
-          {
-            "result": {
-              "duration": 6708470200,
+              "duration": 181004500,
               "status": "passed"
             },
             "line": 12,
-            "name": "user is on home page",
+            "name": "user clicks on login button",
             "match": {
-              "location": "DealsStepDefinition.user_is_on_home_page()"
+              "location": "DealsStepWithMapDefinition.user_clicks_on_login_button()"
             },
             "keyword": "Then "
           },
           {
             "result": {
-              "duration": 4223765000,
+              "duration": 6432565500,
               "status": "passed"
             },
             "line": 13,
-            "name": "user moves to new deal page",
+            "name": "user is on home page",
             "match": {
-              "location": "DealsStepDefinition.user_moves_to_new_deal_page()"
+              "location": "DealsStepWithMapDefinition.user_is_on_home_page()"
             },
             "keyword": "Then "
           },
           {
             "result": {
-              "duration": 7187639000,
+              "duration": 8489590200,
               "status": "passed"
             },
             "line": 14,
+            "name": "user moves to new deal page",
+            "match": {
+              "location": "DealsStepWithMapDefinition.user_moves_to_new_deal_page()"
+            },
+            "keyword": "Then "
+          },
+          {
+            "result": {
+              "duration": 26073277299,
+              "status": "passed"
+            },
+            "line": 15,
             "name": "user enters deal details",
             "match": {
-              "location": "DealsStepDefinition.user_enters_contacts_details(DataTable)"
+              "location": "DealsStepWithMapDefinition.user_enters_contacts_details(DataTable)"
             },
             "rows": [
               {
                 "cells": [
+                  "title",
+                  "amount",
+                  "probability",
+                  "commission"
+                ],
+                "line": 16
+              },
+              {
+                "cells": [
                   "test deal1",
-                  "100",
+                  "10020",
                   "50",
                   "10"
                 ],
-                "line": 15
+                "line": 17
+              },
+              {
+                "cells": [
+                  "test deal2",
+                  "10021",
+                  "55",
+                  "15"
+                ],
+                "line": 18
+              },
+              {
+                "cells": [
+                  "test deal3",
+                  "10022",
+                  "56",
+                  "23"
+                ],
+                "line": 19
               }
             ],
             "keyword": "Then "
           },
           {
             "result": {
-              "duration": 2753718700,
+              "duration": 2269538300,
               "status": "passed"
             },
-            "line": 17,
+            "line": 21,
             "name": "Close the browser",
             "match": {
-              "location": "DealsStepDefinition.close_the_browser()"
+              "location": "DealsStepWithMapDefinition.close_the_browser()"
             },
             "keyword": "Then "
           }
@@ -145,6 +179,6 @@
     "description": "",
     "id": "deal-data-feature",
     "keyword": "Feature",
-    "uri": "D:/Automation/code/FreeCRMBDDFramework/src/main/java/Features/deals.feature"
+    "uri": "D:/Automation/code/FreeCRMBDDFramework/src/main/java/Features/dealsmap.feature"
   }
 ]
