+++ src/main/java/MyRunner/TestRunner.java
@@ -8,7 +8,7 @@ import cucumber.api.junit.Cucumber;
 
     @RunWith(Cucumber.class)
     @CucumberOptions(
-            features = "D:\\Automation\\code\\FreeCRMBDDFramework\\src\\main\\java\\Features\\deals.feature", //the path of the feature files
+            features = "D:\\Automation\\code\\FreeCRMBDDFramework\\src\\main\\java\\Features\\dealsmap.feature", //the path of the feature files
             glue={"stepDefinitions"}, //the path of the step definition files
            format = {"pretty", "html:test-outout", "json:json_output/cucumber.json", "junit:junit_xml/cucumber.xml"} , //to generate different types of reporting
             dryRun = false,
