+++ test-outout/report.js
@@ -1,4 +1,4 @@
-$(document).ready(function() {var formatter = new CucumberHTML.DOMFormatter($('.cucumber-report'));formatter.uri("D:/Automation/code/FreeCRMBDDFramework/src/main/java/Features/deals.feature");
+$(document).ready(function() {var formatter = new CucumberHTML.DOMFormatter($('.cucumber-report'));formatter.uri("D:/Automation/code/FreeCRMBDDFramework/src/main/java/Features/dealsmap.feature");
 formatter.feature({
   "line": 1,
   "name": "Deal Data Feature",
@@ -35,111 +35,145 @@ formatter.step({
   "rows": [
     {
       "cells": [
+        "username",
+        "password"
+      ],
+      "line": 9
+    },
+    {
+      "cells": [
         "Lokeshnp77@gmail.com",
         "F!rew0rk"
       ],
-      "line": 9
+      "line": 10
     }
   ],
-  "keyword": "Then "
-});
-formatter.step({
-  "line": 11,
-  "name": "user clicks on login button",
   "keyword": "Then "
 });
 formatter.step({
   "line": 12,
-  "name": "user is on home page",
+  "name": "user clicks on login button",
   "keyword": "Then "
 });
 formatter.step({
   "line": 13,
-  "name": "user moves to new deal page",
+  "name": "user is on home page",
   "keyword": "Then "
 });
 formatter.step({
   "line": 14,
+  "name": "user moves to new deal page",
+  "keyword": "Then "
+});
+formatter.step({
+  "line": 15,
   "name": "user enters deal details",
   "rows": [
     {
       "cells": [
+        "title",
+        "amount",
+        "probability",
+        "commission"
+      ],
+      "line": 16
+    },
+    {
+      "cells": [
         "test deal1",
-        "100",
+        "10020",
         "50",
         "10"
       ],
-      "line": 15
+      "line": 17
+    },
+    {
+      "cells": [
+        "test deal2",
+        "10021",
+        "55",
+        "15"
+      ],
+      "line": 18
+    },
+    {
+      "cells": [
+        "test deal3",
+        "10022",
+        "56",
+        "23"
+      ],
+      "line": 19
     }
   ],
   "keyword": "Then "
 });
 formatter.step({
-  "line": 17,
+  "line": 21,
   "name": "Close the browser",
   "keyword": "Then "
 });
 formatter.match({
-  "location": "DealsStepDefinition.user_is_already_on_login_page()"
+  "location": "DealsStepWithMapDefinition.user_is_already_on_login_page()"
 });
 formatter.result({
-  "duration": 34689538500,
+  "duration": 24092115700,
   "status": "passed"
 });
 formatter.match({
-  "location": "DealsStepDefinition.title_of_login_page_is_free_CRM()"
+  "location": "DealsStepWithMapDefinition.title_of_login_page_is_free_CRM()"
 });
 formatter.result({
-  "duration": 149732200,
+  "duration": 90695400,
   "status": "passed"
 });
 formatter.match({
-  "location": "DealsStepDefinition.user_clicks_on_login_button_first_time()"
+  "location": "DealsStepWithMapDefinition.user_clicks_on_login_button_first_time()"
 });
 formatter.result({
-  "duration": 20392669900,
+  "duration": 18680433501,
   "status": "passed"
 });
 formatter.match({
-  "location": "DealsStepDefinition.user_enters_username_and_password(DataTable)"
+  "location": "DealsStepWithMapDefinition.user_enters_username_and_password(DataTable)"
 });
 formatter.result({
-  "duration": 1434677900,
+  "duration": 2605164200,
   "status": "passed"
 });
 formatter.match({
-  "location": "DealsStepDefinition.user_clicks_on_login_button()"
+  "location": "DealsStepWithMapDefinition.user_clicks_on_login_button()"
 });
 formatter.result({
-  "duration": 173256300,
+  "duration": 181004500,
   "status": "passed"
 });
 formatter.match({
-  "location": "DealsStepDefinition.user_is_on_home_page()"
+  "location": "DealsStepWithMapDefinition.user_is_on_home_page()"
 });
 formatter.result({
-  "duration": 6708470200,
+  "duration": 6432565500,
   "status": "passed"
 });
 formatter.match({
-  "location": "DealsStepDefinition.user_moves_to_new_deal_page()"
+  "location": "DealsStepWithMapDefinition.user_moves_to_new_deal_page()"
 });
 formatter.result({
-  "duration": 4223765000,
+  "duration": 8489590200,
   "status": "passed"
 });
 formatter.match({
-  "location": "DealsStepDefinition.user_enters_contacts_details(DataTable)"
+  "location": "DealsStepWithMapDefinition.user_enters_contacts_details(DataTable)"
 });
 formatter.result({
-  "duration": 7187639000,
+  "duration": 26073277299,
   "status": "passed"
 });
 formatter.match({
-  "location": "DealsStepDefinition.close_the_browser()"
+  "location": "DealsStepWithMapDefinition.close_the_browser()"
 });
 formatter.result({
-  "duration": 2753718700,
+  "duration": 2269538300,
   "status": "passed"
 });
 });
